+++ nisystemlink/clients/testmonitor/models/__init__.py
@@ -1,6 +1,6 @@
 from ._api_info import Operation, V2Operations, ApiInfo
 from ._product import Product
-from ._create_products_request import CreateProductsPartialSuccess
+from ._create_products_partial_success import CreateProductsPartialSuccess
 from ._delete_products_partial_success import DeleteProductsPartialSuccess
 from ._paged_products import PagedProducts
 from ._query_products_request import (
